remove adder, hadnle in FileNode
