+++ src/Testing/TestingParser/TestingParser.php
@@ -7,6 +7,8 @@ namespace Rector\Testing\TestingParser;
 use Nette\Utils\FileSystem;
 use PhpParser\Node;
 use Rector\Application\Provider\CurrentFileProvider;
+use Rector\CodingStyle\ClassNameImport\UsedImportsResolver;
+use Rector\CodingStyle\ClassNameImport\ValueObject\UsedImports;
 use Rector\NodeTypeResolver\NodeScopeAndMetadataDecorator;
 use Rector\NodeTypeResolver\Reflection\BetterReflection\SourceLocatorProvider\DynamicSourceLocatorProvider;
 use Rector\PhpParser\Node\FileNode;
@@ -23,6 +25,7 @@ final readonly class TestingParser
         private NodeScopeAndMetadataDecorator $nodeScopeAndMetadataDecorator,
         private CurrentFileProvider $currentFileProvider,
         private DynamicSourceLocatorProvider $dynamicSourceLocatorProvider,
+        private UsedImportsResolver $usedImportsResolver,
     ) {
     }
 
@@ -54,8 +57,14 @@ final readonly class TestingParser
         $stmts = $this->rectorParser->parseString($fileContent);
 
         // wrap in FileNode to enable file-level rules
-        $stmts = [new FileNode($stmts)];
+        $stmts = [new FileNode($stmts, new UsedImports([], [], []))];
         $stmts = $this->nodeScopeAndMetadataDecorator->decorateNodesFromFile($filePath, $stmts);
+
+        // seed used imports once, after decoration when namespaced names are resolvable
+        $fileNode = $stmts[0] ?? null;
+        if ($fileNode instanceof FileNode) {
+            $fileNode->setUsedImports($this->usedImportsResolver->resolveForStmts($fileNode->stmts));
+        }
 
         $file->hydrateStmtsAndTokens($stmts, $stmts, []);
         $this->currentFileProvider->setFile($file);
